# apps/server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stationary Server API",
    "description": "All API endpoints for the Stationary backend service, supporting export to tools like Apifox / Postman.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9400",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/tag/list": {
      "get": {
        "summary": "List all tags under the media library (with stats and aliases)",
        "tags": [
          "Tag"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "normalized_name": {
                            "type": "string"
                          },
                          "color": {
                            "type": "string",
                            "nullable": true
                          },
                          "status": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          },
                          "canonical_tag_id": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "post_count": {
                            "type": "integer"
                          },
                          "media_count": {
                            "type": "integer"
                          },
                          "aliases": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ACTIVE",
                "CANDIDATE",
                "IGNORED"
              ],
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/tag/create": {
      "post": {
        "summary": "Create a new tag",
        "tags": [
          "Tag"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "library_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "color": {
                    "type": "string"
                  }
                },
                "required": [
                  "library_id",
                  "name"
                ]
              }
            }
          }
        }
      }
    },
    "/api/tag/update": {
      "post": {
        "summary": "Update tag attributes",
        "tags": [
          "Tag"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "color": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "enum": [
                      "ACTIVE",
                      "CANDIDATE",
                      "IGNORED"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        }
      }
    },
    "/api/tag/delete/{id}": {
      "post": {
        "summary": "Delete tag",
        "tags": [
          "Tag"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tagDeleted": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/tag/merge": {
      "post": {
        "summary": "Merge two tags",
        "tags": [
          "Tag"
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0,
                      "description": "Status code, 0 indicates success"
                    },
                    "message": {
                      "type": "string",
                      "example": "success",
                      "description": "Status description"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 40101
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "library_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "source_tag_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "target_tag_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "retain_as_alias": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "library_id",
                  "source_tag_id",
                  "target_tag_id"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authenticate requests using API token (Authorization: Bearer <token>)"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session-token",
        "description": "Authenticate requests using Better Auth session cookie"
      }
    }
  }
}